back_end/requirements.txt
Flask==2.3.2
flask-cors==3.1.1
numpy==1.24.2
Pillow==9.4.0
tensorflow==2.13.0
